+++ src/ext/Bootstrap/Bootstrap_helper_registry.cpp
@@ -12,7 +12,7 @@ void RegisterIDRegistry() {
 	Mngr->AddMethod<MemFuncOf<W, std::string_view(T, std::string_view)>::get(&W::Register)>("Register");
 	Mngr->AddMethod<MemFuncOf<W, U(std::string_view)>::get(&W::Register)>("Register");
 	Mngr->AddMethod<&W::IsRegistered>("IsRegistered");
-	Mngr->AddMethod<&W::Nameof>("Nameof");
+	Mngr->AddMethod<&W::Viewof>("Viewof");
 	Mngr->AddMethod<&W::UnregisterUnmanaged>("UnregisterUnmanaged");
 	Mngr->AddMethod<&W::Clear>("Clear");
 }
@@ -20,11 +20,15 @@ void RegisterIDRegistry() {
 void Ubpa::UDRefl::ext::details::Bootstrap_helper_registry() {
 	RegisterIDRegistry<NameID, Name>();
 	Mngr->RegisterType<NameIDRegistry>();
+	Mngr->AddMethod<MemFuncOf<NameIDRegistry, Name(Name)>::get(&NameIDRegistry::Register)>("Register");
+	Mngr->AddMethod<&NameIDRegistry::Nameof>("Nameof");
 	Mngr->AddBases<NameIDRegistry, IDRegistry<NameID, Name>>();
 
 	RegisterIDRegistry<TypeID, Type>();
 	Mngr->RegisterType<TypeIDRegistry>();
 	Mngr->AddBases<TypeIDRegistry, IDRegistry<TypeID, Type>>();
+	Mngr->AddMethod<MemFuncOf<TypeIDRegistry, Type(Type)>::get(&TypeIDRegistry::Register)>("Register");
+	Mngr->AddMethod<&TypeIDRegistry::Typeof>("Typeof");
 	Mngr->AddMethod<&TypeIDRegistry::RegisterAddConst>("RegisterAddConst");
 	Mngr->AddMethod<&TypeIDRegistry::RegisterAddLValueReference>("RegisterAddLValueReference");
 	Mngr->AddMethod<&TypeIDRegistry::RegisterAddLValueReferenceWeak>("RegisterAddLValueReferenceWeak");